+++ docs/en/changes/changes.md
@@ -31,6 +31,7 @@
 * Fix expression of graph `Current QPS` in MySQL dashboard.
 * Support tracing logs query for debugging.
 * BanyanDB: fix Tag autocomplete data storage and query.
+* Support aggregation operators in PromQL query.
 
 #### UI
 * Highlight search log keywords.